// app/data/remote/Auth/model/response/AuthResponse.ts
export interface IAuthResp {
  courier_id:              string;
  courier_name:            string;
  courier_password:        string;
  courier_phone:           null;
  courier_email:           null;
  courier_nik:             null;
  courier_regional:        string;
  courier_branch:          string;
  courier_zone:            string;
  courier_origin:          string;
  courier_active:          number;
  courier_sp_value:        string;
  courier_incentive_group: string;
  courier_armada:          string;
  courier_employee_status: null;
  courier_created_at:      string;
  courier_updated_at:      string;
  courier_role_id:         string;
  courier_level:           null;
  team_id:                 null;
  verified:                boolean;
  access_token:            string;
}